Convert timeStamp property of Message class back to Long so that we can format it as we need.

+++ app/src/main/java/com/ndhunju/relay/util/Utils.kt
@@ -21,7 +21,7 @@ fun readSms(contentResolver: ContentResolver): List<Message> {
             val message = Message(
                 cursor.getStringForColumn("address"),
                 cursor.getStringForColumn("body"),
-                cursor.getStringForColumn("date")
+                cursor.getStringForColumn("date").toLong()
             )
             //println("SMS from: $message")
             messages.add(message)
